+++ build_test/lib/src/resolve_source.dart
@@ -50,21 +50,66 @@ import 'package_reader.dart';
 /// - by default, [PackageAssetReader.currentIsolate]. A custom [resolver] may
 /// be provided to map files not visible to the current package's runtime.
 Future<Resolver> resolveSource(String inputSource,
-    {AssetId inputId, PackageResolver resolver, Future<Null> tearDown}) async {
-  // If not provided use a fake asset and package.
-  inputId ??= new AssetId('_resolve_source', '_resolve_source.dart');
+        {AssetId inputId, PackageResolver resolver, Future<Null> tearDown}) =>
+    _resolveAsset(
+        inputId ?? new AssetId('_resolver_source', 'lib/_resolve_source.dart'),
+        inputContents: inputSource,
+        resolver: resolver,
+        tearDown: tearDown);
+
+/// Returns a future that completes with a source [Resolver] for [input].
+///
+/// Example use:
+/// ```dart
+/// var pkgBuildTest = new AssetId('build_test', 'lib/build_test.dart');
+/// var resolver = await resolveSource(pkgBuildTest);
+/// ```
+///
+/// By default, the returned [Resolver] is destroyed after the event loop is
+/// completed. In order to control the lifecycle, pass a [Completer] and
+/// complete it turning the `tearDown` phase of testing:
+/// ```dart
+/// Completer<Null> onTearDown;
+///
+/// setUp(() {
+///   onTearDown = new Completer<Null>();
+/// });
+///
+/// tearDown(() => onTearDown.complete());
+///
+/// test('...', () async {
+///   var resolver = await resolveAsset('...', tearDown: onTearDown.future);
+///   // Use resolver.
+/// });
+/// ```
+///
+/// **NOTE**: All `package` dependencies are resolved using [PackageAssetReader]
+/// - by default, [PackageAssetReader.currentIsolate]. A custom [resolver] may
+/// be provided to map files not visible to the current package's runtime.
+Future<Resolver> resolveAsset(AssetId input,
+        {PackageResolver resolver, Future<Null> tearDown}) =>
+    _resolveAsset(input, resolver: resolver, tearDown: tearDown);
+
+/// Internal only backing implementation of `resolveAsset` and `resolveSource`.
+///
+/// If [inputContents] is non-null, it is used instead of reading [input] from
+/// the file system.
+Future<Resolver> _resolveAsset(AssetId input,
+    {String inputContents,
+    PackageResolver resolver,
+    Future<Null> tearDown}) async {
   resolver ??= PackageResolver.current;
   tearDown ??= new Future.delayed(Duration.ZERO);
   var syncResolver = await resolver.asSync;
-  var reader = new PackageAssetReader(syncResolver, inputId.package);
+  var reader = new PackageAssetReader(syncResolver, input.package);
   var completer = new Completer<Resolver>();
   var builder = new _ResolveSourceBuilder(completer, tearDown);
-  var inputs = [inputId];
+  var inputs = [input];
   var inMemory = new InMemoryAssetReader(
     sourceAssets: {
-      inputId: new DatedString(inputSource),
+      input: new DatedString(inputContents ?? await reader.readAsString(input)),
     },
-    rootPackage: inputId.package,
+    rootPackage: input.package,
   );
   // We don't care about the results of this build.
   // ignore: unawaited_futures
@@ -74,7 +119,7 @@ Future<Resolver> resolveSource(String inputSource,
     new MultiAssetReader([inMemory, reader]),
     new InMemoryAssetWriter(),
     const BarbackResolvers(),
-    rootPackage: inputId.package,
+    rootPackage: input.package,
   );
   return completer.future;
 }